refactor : money값들 String에서 Long으로 변경(BIGINT)

+++ src/main/java/com/oopsw/clario/dto/dashboard/TradeDTO.java
@@ -10,9 +10,9 @@ import lombok.*;
 public class TradeDTO {
     private int memberId;
     private String cardStoreName;
-    private int cardTradeMoney;
+    private Long cardTradeMoney;
     private String accountSource;
-    private int accountTradeMoney;
+    private Long accountTradeMoney;
     private String cardTradeMonth;
     private String accountTradeMonth;
 }
